fix(permissios): Once and for all makes new version working (#7)
Revert 69210242c702bdaa1f53f22287f859a12aaacbc9 and use `root` as
container user. The reason for this ugly change is that all Github
Workflows (one for each OF repository) are coupled with this
requirement: they expect a to find a `gitleaks-output.json` file in the
root folder of the container (i.e.
    [here](https://github.com/motain/android-iLiga-Libraries/actions/runs/1819251186/workflow#L4))

This comes again under the umbrella of technical debts.